requirements.txt
opencv-python
mediapipe
numpy
pillow
pyvirtualcam
